requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
python-slugify==1.2.6
requests==2.20.0
selenium==3.14.1
Unidecode==1.0.22
urllib3==1.24
